+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 2.2.0
+- Add `[DateTime]` support to `FREArray`
+- Change: `FREArray` now uses Vector.<Type> throughout
+- Improve `ANEUtils.map` performance
+- Upgrade to AIR 32.0.0.116
+
 ### 2.1.0
 - Upgraded to AIR 32
 - Added `FREArray.IsEmpty`